Introduce play framework as another backend option
This required multiple changes. I had to downgrade the java version to
11, so I'm using jenv to manage different java versions. Also had to
downgrade sbt to a lower version due to dependency conflicts.
Finally, I realized that sbt-tpolecat was hijacking scalac options
which is fine...just didn't realize that and was really confused
as to why me adding certain compiler options was not doing anything.