Use zio http and add cors
Well this required a lot of tinkering...
Lots of little api changes due to using a different verison of zio http
(zio-http vs zhttp)
In addition to that, I had to add orDie to every route since the serve
method from Server requires an App type, which is an
HttpApp[R, Response], meaning
there can't be any throwables.
Alternatively, I could've also
changed the UrlRepo to use Nothing instead
of Throwable, but I still had to add orDie
to the POST route since asString
returns a Task which contains Throwable.

So...lots of issues with this. The quickstarts from ZIO and the docs are
a little inconsistent, I may try updating them later.